+++ src/components/SearchBar/index.js
@@ -1,0 +1,2 @@
+export { Logo } from "./Logo";
+export { SearchBar } from "./SearchBar";
