+++ src/components/SearchResults/index.js
@@ -1,5 +1,3 @@
-export { Logo } from "./Logo";
-export { SearchBar } from "./SearchBar";
 export { SearchResults } from "./SearchResults";
 export { Forks } from "./Forks";
 export { Files } from "./Files";
